### 2023-08-29 #131 @remytuyeras
Presubmission Inquiry for HaploDynamics
Submitting Author: Remy Tuyeras (@remytuyeras)
Package Name: HaploDynamics
One-Line Description of Package:  A python library to develop genomic data simulators
Repository Link (if existing):  https://github.com/remytuyeras/HaploDynamics

---

The **HaploDynamics package** provides a collection of functions to generate simulated population-specific genomic data in VCF format. The library includes parameters and functions to control mutation rates, linkage disequilibrium strength and block lengths, and number of individuals. To generate genomic data, the HaploDX framework offers a pipeline of functions that can be used to simulate: (1) the allele frequency spectra of different populations; (2) the Hardy-Weinberg principle for genotypes and haplotypes; (3) linkage disequilibrium across different populations.

We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an

I am not sure whether the package falls under one of the categories and was wondering if you editorial team could guide me on this.

The decreasing costs of DNA sequencing and cloud computing have led to a surge in the deployment of genomic pipelines on cloud platforms for international collaboration. This development has made it increasingly important to develop privacy-preserving methods that protect both patient privacy and the integrity of research that uses the data. Regulatory and privacy-preserving protocols have made genomic pipeline auditing more stringent and thorough, requiring them to be tested on simulated data to verify the validity of the algorithms that process the data. However, the process of iteratively generating and/or transferring genomic simulations on the cloud can be a major bottleneck for research.

To address this challenge, the **HaploDynamics** package provides a cloud-native framework for generating fast and lightweight simulations of genomic data that can replicate its complexity. The package also allows researchers to add their own generative mutation models to the framework, making it possible to create customized simulations that meet specific research needs.

MaCS (coded in C++) is a fast coalescence-based method for generating genomic data. However, its use of graph structures to encode segregation and recombination events can slow down its execution time.

Both MaCS and HaploDX rely on Bayesian networks to generate variants, which allows them to output rows of genotypes sequentially in a similar fashion. However, HaploDX does not use graph structures, relying instead on additional mathematical considerations. This difference allows HaploDX to outperform MaCS on execution times and memory usage.

Other packages, such as Sim1000G (coded in R), also rely on coalescence/pedigree structures, which can impair their execution times (and memory usage).

Ref:
- https://github.com/adimitromanolakis/sim1000G
- https://github.com/gchen98/macs/tree/master

I am wondering if the HaploDynamics package fits the scope of the pyOpenSci project. If not, I would like to know if you have any recommendations for other packages or projects that I should consider.

Further, while HaploDynamics is fully functional, it is also under active development of new functionalities. Is this something that would be compatible with the pyOpenSci project?

[Conduct]: https://www.pyopensci.org/governance/CODE_OF_CONDUCT
