### 2023-08-29 #131 @NickleDave
Welcome to the pyOpenSci community @remytuyeras.

Yes, at first glance, HaploDynamics looks like it is in scope.

What would help us determine this would be complete documentation for the project, including examples that directly use HaploDynamics itself, that are built and rendered as a separate site.

Such documentation would also required before a full submission.

I do see you have documentation in docs/source and that you point to a blog post on your personal site. As far as I can tell the blog post does not specifically show usage of HaploDynamics itself, and is instead a more general tutorial on simulating pop genomics with Python. It's great that you have written that, but for submission you will need examples usages of HaploDynamics in your documentation. I see some example snippets in docs/source but you will also want stand alone examples of "how-tos" or "walkthroughs" such as [this](https://robaina.github.io/Pynteny/examples/example_api/) and [this](https://robaina.github.io/Pynteny/examples/example_sus/) in the pynteny documentation.

What you will need to do is:
- [ ] Set up your docs so they can be rendered as a separate web page by Sphinx or MkDocs
  - [ ] Note that you can mostly avoid rewriting from Markdown to RestructuredText by using MyST, as discussed [here](https://www.pyopensci.org/python-package-guide/documentation/hosting-tools/myst-markdown-rst-doc-syntax.html)
  - [ ] You can probably recycle a large part of your README as a landing page for the docs, and avoid having the two go out of sync, by using a [literal include](https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#directive-literalinclude) of the README in your docs
  - [ ] API documentation -- I see this page but it looks like you are defining the API by hand? You may find it helpful to instead use [autodoc](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html) and [autosummary](https://www.sphinx-doc.org/en/master/usage/extensions/autosummary.html) for example as done [here](https://github.com/GenericMappingTools/pygmt/blob/main/doc/api/index.rst) in the PyGMT API docs
- [ ] build and render the docs with an external service such as ReadTheDocs or GitHub Pages, as described in [this section](https://www.pyopensci.org/python-package-guide/documentation/hosting-tools/publish-documentation-online.html) of our guide

Please let me know if that is clear.
I'll be happy to have a look and get more feedback from our editorial board on whether HaploDynamics is in scope once you have full documentation that will help us better understand the intended audience, functionality, and usage of the package. It looks like you are most of the way there now! Just set up a build please 🙂 and make sure the rendered pages include all the great info that's in your README
