### 2023-08-29 #131 @NickleDave
Hi again @remytuyeras, I forgot to say:
it's not strictly required for submission, but you probably want to replace your setup.py with a pyproject.toml file that explicitly declares your dependencies in the metadata section, including scipy, as discussed here: https://www.pyopensci.org/python-package-guide/package-structure-code/pyproject-toml-python-package-metadata.html

Explicitly declaring your dependencies in your pyproject.toml will allow you to remove the note in your README instructing users to install scipy. Once you state it is a dependency in your pyproject.toml and publish a new version of HaploDynamics to PyPI, your users will get SciPy installed automatically when they `pip install haploydynamics'. Please also check out our guide on build tools here: https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-build-tools.html
