### 2023-08-30 #131 @remytuyeras
Hi NickleDave,

Thank you for your thorough feedback, this is very much appreciated!

I will focus on setting up a documentation and will get back to pyOpenSci when this is done.

I would like to address some of the points made in the second message by providing some complementary information.
- The **SciPy** package is used in the ```HaploDynamics/HaploDX/haploDX.py``` module to compute Pearson correlation scores. This is done by importing the ```scipy.stats``` submodule as ```stat``` and then calling the function ```stat.pearsonr()``` inside the function ```LD_corr_matrix()```.
- I intentionally distinguished the modules ```HaploDX``` and ```Framework```. ```HaploDX``` is an introductory module that is not optimized, but it is easier to start with. It stems from the tutorial published on my personal webpage. Anyone interested in **learning** the framework can experiment with ```HaploDynamics.HaploDX``` while following the tutorial. The module ```Framework``` is specifically optimized for **research** and is intended for researchers who need the most performance. However, both modules have the same underlying principles.

I also have a question regarding the difference between **pyproject.toml** and **setup.py**:
- In the README, my intention was to let the user know that they need to install scipy only if they want to set up the PYTHONPATH variable to the package manually. Otherwise, I included SciPy in setup.py using ```install_requires=['scipy']``` and thought that this inclusion would automatically install the dependency when using ```pip```. After your suggestion, I am wondering whether this is a misconception, or is there any other advantage to using pyproject.toml in the context of pyOpenSci?

Please let me know if you have any feedback given the additional information above. Any tip or advice is appreciated. I'm always looking for ways to improve the package design.
