### 2023-08-30 #131 @NickleDave
Hi again @remytuyeras

> I will focus on setting up a documentation and will get back to pyOpenSci when this is done.

Great. Thank you for working on that. Just so we know where we're at, I added the "pending-maintainer-response" label to this issue.

> The SciPy package is used in the HaploDynamics/HaploDX/haploDX.py module
> Otherwise, I included SciPy in setup.py using install_requires=['scipy'] and thought that this inclusion would automatically install the dependency when using pip

I'm sorry, it's my fault for not reading the setup.py more closely. I tested that I could `pip install HaploDynamics` and I did see that `scipy` was installed and I could successfully import `HaploDynamics`.

> I intentionally distinguished the modules HaploDX and Framework.

This is a really great design choice, and I think it's very helpful to provide easy-to-read pure Python implementations as well as more efficient implementations that e.g. are vectorized with NumPy.

I also see now that you *do* import these modules at the package level inside `HaploDynamics.__init__.py` -- again, my fault for not reading more carefully.

That said, I would **strongly** suggest you use lowercase for both the package and module names, to avoid confusing people reading code that uses HaploDynamics, who will mistake them for classes because you have used the convention for naming classes. Please see the PEP8 conventions for [package and module names](https://peps.python.org/pep-0008/#package-and-module-names) and [class names](https://peps.python.org/pep-0008/#class-names). Of course we know that ["a foolish consistency is the hobgoblin of small minds"](https://peps.python.org/pep-0008/#a-foolish-consistency-is-the-hobgoblin-of-little-minds) but here we want to be consistent so people can easily read all the code that's going to be using your library! You don't want to make all those people mad at you 🙂

> is there any other advantage to using pyproject.toml

There are many advantages.
The main one is that it [lets you declare what tools to use to build your package](https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/), in static metadata instead of code. It also lets you declare additional metadata associated with the package, that [other tools](https://snarky.ca/what-the-heck-is-pyproject-toml/) can [take advantage of](https://github.blog/changelog/2023-02-13-dependency-graph-supports-the-python-pep-621-standard/). Most other modern languages use static files for this purpose instead of code. Avoiding the use of setup.py also prevents an entire class of security vulnerabilities, since we are no longer executing code in setup.py when building a package for installation. The only reason you may need a setup.py is if your package is not pure Python and has a more complex build--in your case you only have two dependencies, numpy and scipy, for which wheels are readily available, so you do not need a setup.py file.

Does that help?
