### 2023-08-30 #131 @remytuyeras
Hi @NickleDave ,

Yes, it helps a lot. I will work on the documentation and integrate your feedback above.

Thanks!

Best,
Remy
