### 2023-09-13 #131 @NickleDave
@remytuyeras I am going to go ahead and closed this since we have told you that, yes, HaploDynamics is in scope.

Please do reference this issue when you make the full submission (there's a place to do so in the template you'll fill out). Thank you!
